moved Commands/Strategies to a top level directory, various cleanup and StakeHbar to Facade
