+++ packages/hedera-wallet-snap/packages/snap/src/commands/AssociateTokensCommand.ts
@@ -1,7 +1,7 @@
 import { AccountId, Client, TokenAssociateTransaction } from '@hashgraph/sdk';
-import { TxReceipt } from '../../types/hedera';
-import { Utils } from '../../utils/Utils';
-import { CryptoUtils } from '../../utils/CryptoUtils';
+import { TxReceipt } from '../types/hedera';
+import { Utils } from '../utils/Utils';
+import { CryptoUtils } from '../utils/CryptoUtils';
 
 export class AssociateTokensCommand {
   readonly #tokenIds: string[];
