+++ packages/hedera-wallet-snap/packages/snap/src/commands/DeleteAccountCommand.ts
@@ -1,7 +1,7 @@
 import { AccountDeleteTransaction, Client } from '@hashgraph/sdk';
-import { TxReceipt } from '../../types/hedera';
-import { Utils } from '../../utils/Utils';
-import { CryptoUtils } from '../../utils/CryptoUtils';
+import { TxReceipt } from '../types/hedera';
+import { Utils } from '../utils/Utils';
+import { CryptoUtils } from '../utils/CryptoUtils';
 
 export class DeleteAccountCommand {
   readonly #transferAccountId: string;
